% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_geo.R
\name{get_dataset_summary}
\alias{get_dataset_summary}
\title{Get Dataset Summary Information}
\usage{
get_dataset_summary(eset)
}
\arguments{
\item{eset}{An ExpressionSet object.}
}
\value{
A list with:
  - n_features: Number of genes/probes
  - n_samples: Number of samples
  - phenotype_cols: Column names in phenotype data
  - expression_range: Min and max expression values
}
\description{
Returns basic summary statistics for a dataset.
}
\examples{
\dontrun{
  eset <- fetch_geo_dataset("GSE63310")
  summary_info <- get_dataset_summary(eset)
  str(summary_info)
}

}
