% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{GSEexplore-package}
\alias{GSEexplore}
\alias{GSEexplore-package}
\title{GSEexplore Package}
\description{
GSEexplore provides a user-friendly interface for:
  - Fetching GEO datasets by accession ID
  - Quality control assessment (PCA, sample distributions)
  - Batch effect detection and correction
  - Heatmap visualization of expression patterns
  - Differential expression analysis using limma
  - Interactive data exploration
}
\details{
An interactive Shiny application for exploring Gene Expression Omnibus (GEO)
datasets with quality control and differential expression analysis capabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/user/GSEexplore}
  \item Report bugs at \url{https://github.com/user/GSEexplore/issues}
}

}
\author{
\strong{Maintainer}: GSEexplore Contributors \email{info@gseexplore.dev}

}
\keyword{internal}
